requirements.txt
streamlit
pypdf
google.generativeai
langchain
python-dotenv